requirements.txt
easyocr==1.7.1
fastapi==0.109.0
Pillow==10.2.0
